backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask_restx
flask_sqlalchemy
flask-session
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2